+++ src/components/StakeCapital/StakeItem.tsx
@@ -1,6 +1,11 @@
-import { Box, TableCell, TableRow, Tooltip, Typography, useTheme } from "@mui/material";
+import {
+  TableCell,
+  TableRow,
+  Tooltip,
+  Typography,
+  useTheme,
+} from "@mui/material";
 import { CSSProperties, useEffect, useState } from "react";
-import { useQuery } from "react-query";
 import { AccountInterface } from "starknet";
 
 import { getPoolState } from "../../calls/getPoolState";
@@ -8,15 +13,13 @@ import { Pool } from "../../classes/Pool";
 import { TokenKey } from "../../classes/Token";
 import { BASE_DIGITS } from "../../constants/amm";
 import { useTxPending } from "../../hooks/useRecentTxs";
-import { QueryKeys } from "../../queries/keys";
 import { TransactionAction } from "../../redux/reducers/transactions";
 import buttonStyles from "../../style/button.module.css";
 import inputStyle from "../../style/input.module.css";
 import { handleNumericChangeFactory } from "../../utils/inputHandling";
 import { intToDecimal } from "../../utils/units";
-import { LoadingAnimation } from "../Loading/Loading";
-import { fetchStakeCapital } from "./fetchStakeCapital";
 import { handleStake } from "./handleStake";
+import { CapitalItem } from "./CapitalItem";
 
 type Props = {
   account: AccountInterface | undefined;
@@ -69,25 +72,15 @@ export const StakeCapitalItem = ({ account, pool }: Props) => {
   const [text, setText] = useState<string>("0");
   const [loading, setLoading] = useState<boolean>(false);
   const [apy, setApy] = useState<number | undefined>();
-  const [unCapital,setUnCapital] = useState("0");
-  const [lockCapital,setLockCapital] = useState("0");
   const [yieldSinceLaunch, setYieldSinceLaunch] = useState<
     number | undefined
   >();
-  const { isLoading, isError, data } = useQuery(
-    [QueryKeys.stake, pool.name],
-    fetchStakeCapital
-  );
   const theme = useTheme();
 
   useEffect(() => {
     getApy(setApy, pool);
     getYieldSinceLaunch(setYieldSinceLaunch, pool);
-    if(data && data.status==="success"){
-      setUnCapital(BigInt(data.data.unlocked_cap).toString(10));
-      setLockCapital(BigInt(data.data.locked_cap).toString(10));
-    }
-  }, [isLoading, data, pool]);
+  }, [pool]);
 
   const handleChange = handleNumericChangeFactory(setText, setAmount);
 
@@ -160,22 +153,7 @@ export const StakeCapitalItem = ({ account, pool }: Props) => {
           </button>
         </TableCell>
       </TableRow>
-      {showLockInfo && (
-        <TableRow>
-          {isLoading ? (
-            <Box sx={{ padding: "20px" }}>
-              <LoadingAnimation size={20} />
-            </Box>
-          ) : (
-            <>
-              {" "}
-              <TableCell>UNLOCKED CAPITAL: {unCapital}</TableCell>
-              <TableCell>LOCKED CAPITAL: {lockCapital}</TableCell>
-            </>
-          )}
-          {isError&&(<TableCell>Failed fetching data.</TableCell>)}
-        </TableRow>
-      )}
+      {showLockInfo && <CapitalItem pool={pool} />}
     </>
   );
 };
